+++ packages/core/src/scuttle.d.ts
@@ -1,25 +1,11 @@
-/**
- * Options for scuttling global properties
- */
-export type LavaMoatScuttleOpts = {
-  /** Whether scuttling is enabled or not. */
-  enabled: boolean;
-  /** List of properties to exclude from scuttling. */
-  exceptions?: Array<string | RegExp>;
-  /** Name of the scuttler function to use which is expected to be found as a property on the global object (e.g. if scuttlerName is 'x', scuttler function is obtained from globalThis['x']). */
-  scuttlerName?: string;
-}
+// Declaring types here and importing them in scuttle.js 
+// leads to typescript getting confused into generating 
+// a self-referential cycle in the generated .d.ts files:
+// types/src/scuttle.d.ts
+// import type { LavaMoatScuttleOpts } from './scuttle.d.ts';
 
-/**
- * Reference to the global object
- */
-export interface GlobalRef {
-  globalThis?: Record<PropertyKey, unknown>;
-}
+export type {
+    LavaMoatScuttleOpts
+} from './scuttle.js'
 
-/**
- * Scuttles (disables or restricts) certain global properties
- * @param globalRef The global object to scuttle
- * @param [opts] Options for scuttling
- */
-export function scuttle(globalRef: GlobalRef, opts?: LavaMoatScuttleOpts | boolean): void;
+export { scuttle } from './scuttle.js'