+++ packages/lavapack/CHANGELOG.md
@@ -24,6 +24,69 @@
   * dependencies
     * lavamoat-core bumped from ^15.2.0 to ^15.2.1
 
+## [7.0.21](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.20...lavapack-v7.0.21) (2025-12-02)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^17.1.1 to ^17.1.2
+
+## [7.0.20](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.19...lavapack-v7.0.20) (2025-12-01)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^17.1.0 to ^17.1.1
+
+## [7.0.19](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.18...lavapack-v7.0.19) (2025-11-19)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^17.0.1 to ^17.1.0
+
+## [7.0.18](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.17...lavapack-v7.0.18) (2025-10-30)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^17.0.0 to ^17.0.1
+
+## [7.0.17](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.16...lavapack-v7.0.17) (2025-10-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^16.7.1 to ^17.0.0
+
+## [7.0.16](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.15...lavapack-v7.0.16) (2025-10-01)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^16.7.0 to ^16.7.1
+
+## [7.0.15](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.14...lavapack-v7.0.15) (2025-09-25)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * lavamoat-core bumped from ^16.6.2 to ^16.7.0
+
 ## [7.0.14](https://github.com/LavaMoat/LavaMoat/compare/lavapack-v7.0.13...lavapack-v7.0.14) (2025-09-22)